Added Roll Value extractor
The new version allows C++ to send an integer value on any buffer that
can then be extracted in Pd. After being extracted the C++ output is
reformed in Pd so that the file may be played. However, this needs
reworking as the unpackOSC seems to have trouble but I believe this is
purely due to a miss written mock piece of ascii